+++ addons/AsepriteWizard/interface/docks/animated_sprite/inspector_plugin.gd
@@ -3,9 +3,6 @@ extends EditorInspectorPlugin
 
 const ASInspectorDock = preload("./animated_sprite_inspector_dock.tscn")
 
-var config
-var file_system: EditorFileSystem
-
 func _can_handle(object):
 	return object is AnimatedSprite2D || object is AnimatedSprite3D
 
@@ -13,6 +10,4 @@ func _can_handle(object):
 func _parse_end(object):
 	var dock = ASInspectorDock.instantiate()
 	dock.target_node = object
-	dock.config = config
-	dock.file_system = file_system
 	add_custom_control(dock)