+++ addons/AsepriteWizard/interface/docks/sprite/inspector_plugin.gd
@@ -3,8 +3,6 @@ extends EditorInspectorPlugin
 
 const APInspectorDock = preload("./sprite_inspector_dock.tscn")
 
-var config
-var file_system: EditorFileSystem
 
 func _can_handle(object):
 	return object is Sprite2D || object is Sprite3D || object is TextureRect
@@ -13,6 +11,4 @@ func _can_handle(object):
 func _parse_end(object):
 	var dock = APInspectorDock.instantiate()
 	dock.target_node = object
-	dock.config = config
-	dock.file_system = file_system
 	add_custom_control(dock)