+++ addons/AsepriteWizard/interface/docks/wizard/as_wizard_dock_container.gd
@@ -5,26 +5,22 @@ signal close_requested
 
 const WizardWindow = preload("./as_wizard_window.tscn")
 
-var _config
-var _file_system: EditorFileSystem
-
 func _ready():
-	$Import.init(_config, _file_system)
-	$Import.connect("close_requested",Callable(self,"emit_signal").bind("close_requested"))
-	$Import.connect("import_success",Callable($History,"add_entry"))
-	$History.init(_config)
-	$History.connect("request_edit",Callable(self,"_on_edit_request"))
-	$History.connect("request_import",Callable(self,"_on_import_request"))
+	$Import.close_requested.connect(emit_signal.bind("close_requested"))
+	$Import.import_success.connect(_on_import_success)
+	$History.request_edit.connect(_on_edit_request)
+	$History.request_import.connect(_on_import_request)
+	$ImportedSpriteFrames.import_success.connect($History.add_entry)
 
-
-func init(config, editor_file_system: EditorFileSystem):
-	_config = config
-	_file_system = editor_file_system
+	self.set_tab_title(1, "Imported Resources")
 
 
 func _on_AsWizardDockContainer_tab_changed(tab: int):
-	if tab == 1:
-		$History.reload()
+	match tab:
+		1:
+			$ImportedSpriteFrames.init_resources()
+		2:
+			$History.reload()
 
 
 func _on_edit_request(import_cfg: Dictionary):
@@ -35,3 +31,9 @@ func _on_edit_request(import_cfg: Dictionary):
 func _on_import_request(import_cfg: Dictionary):
 	$Import.load_import_config(import_cfg)
 	$Import.trigger_import()
+
+
+func _on_import_success(settings: Dictionary):
+	$ImportedSpriteFrames.init_resources()
+	$ImportedSpriteFrames.reload_tree()
+	$History.add_entry(settings)